Recitation 2 Problem 1a
Proof by ordering, still need to rewrite it as a proof by induction.

Another big todo I noticed while writing this question up is markdown's
lack of mathematical notation. Some of the algebra would be a lot
clearer if I could properly represent fractions, instead of having to
use parentheses for everything.

I'm going to take a look at hosting this on github pages, and using
LaTeX for the doc generation. Good opportunity to learn some
documentation coding on the side!