Determine graphviz location with python
run command stdout doesn't seem to stick around inside github's
workflows console, so even if there was output for the echo command, it
it gets removed immediately.

Since the sphinx-build part (inside python) is the last step to fail,
the stdout remains on the actions console, so I can actually see what
I'm working with.