Refactor questions&introduction with admonitions
I think this looks clearer than bold, and it's a lot easier to type up.
I should probably go back and do this for all the previous work, but I'm
not sure it's really worth it... I don't think anyone will ever read
this, it's mostly for my own benefit.